% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-preprocessing-functions.R
\name{clean}
\alias{clean}
\title{Cleaning of data}
\usage{
clean(
  TE,
  assay = "norm",
  qc_LOQ = "loq",
  new_assay = "clean",
  thres_LOQ = 1.5,
  max_nafrac_per_met = 0.9,
  max_nafrac_per_group = 1/3,
  min_rep_per_group = 2,
  split_by = ~1,
  type = "iso",
  exclude = "internal.standard",
  soft = NULL,
  remove_imp = FALSE,
  ...
)
}
\arguments{
\item{TE}{TracerExperiment}

\item{assay}{norm}

\item{qc_LOQ}{LOQ values}

\item{new_assay}{clean}

\item{thres_LOQ}{LOQ threshold (values below are set to NA)}

\item{max_nafrac_per_group}{maximum allowed fraction of NA values per group}

\item{min_rep_per_group}{minimum required number of non-NA values per group}

\item{split_by}{factor(s) from colData to define groups}

\item{exclude}{internal standard}

\item{soft}{}

\item{remove_imp}{optionally, remove all imputed values}
}
\description{
Cleaning of data
}
\examples{
exampleTracerExperiment(add_qc = TRUE) |> estimateLOQs() |> clean(assay = "raw")
exampleTracerExperiment(add_qc = TRUE) |> estimateLOQs() |> sumMets(assay = "raw", sum_qc = TRUE) |> clean(assay = "raw", type = "met", qc_LOQ = "met_loq_raw")
}
