% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-isocorrection-functions.R
\name{correctIso}
\alias{correctIso}
\title{Natural isotope abundance correction}
\usage{
correctIso(TE, assay = "imp", ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Natural isotope abundance correction
}
