% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-isocorrection-functions.R
\name{.add1}
\alias{.add1}
\title{Helper function for formatting molecule formulas for IsoCorrector}
\usage{
.add1(tmp)
}
\arguments{
\item{tmp}{}
}
\description{
Helper function for formatting molecule formulas for IsoCorrector
}
