% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-data-functions.R
\name{.getAssays}
\alias{.getAssays}
\title{Function to get default assay}
\usage{
.getAssays(
  TE,
  type = "iso",
  last = FALSE,
  assay = NULL,
  not = NULL,
  names_only = FALSE,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Function to get default assay
}
