% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-isocorrection-functions.R
\name{.makeElementsTable}
\alias{.makeElementsTable}
\title{Internal function to generate elements table for IsoCorrector}
\usage{
.makeElementsTable(df, isocomp, tracer, tracer_purity)
}
\arguments{
\item{tracer_purity}{}
}
\description{
Internal function to generate elements table for IsoCorrector
}
