% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-isocorrection-functions.R
\name{.makeMoleculeTable}
\alias{.makeMoleculeTable}
\title{Internal function to generate molecule table for IsoCorrector}
\usage{
.makeMoleculeTable(molecules, Molecule, MSion, mode, tracer)
}
\arguments{
\item{tracer}{}
}
\description{
Internal function to generate molecule table for IsoCorrector
}
