% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-data-functions.R
\name{.orderLabels}
\alias{.orderLabels}
\title{Order isotopologue labels numerically}
\usage{
.orderLabels(x, pattern = "m")
}
\arguments{
\item{pattern}{}
}
\value{

}
\description{
Order isotopologue labels numerically
}
