% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-data-functions.R
\name{.stretch}
\alias{.stretch}
\title{Simple pivot_longer wrapper for numeric matrices}
\usage{
.stretch(mat)
}
\arguments{
\item{mat}{}
}
\value{

}
\description{
Simple pivot_longer wrapper for numeric matrices
}
