% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-preprocessing-functions.R
\name{estimateLODs}
\alias{estimateLODs}
\title{Estimate LOD/LOQ values from QC samples}
\usage{
estimateLODs(TE, qc = "blanks", exclude = "internal.standard", FUN = NULL, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Estimate LOD/LOQ values from QC samples
}
