% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-preprocessing-functions.R
\name{estimateLOQs}
\alias{estimateLOQs}
\title{Estimate LOD/LOQ values from QC samples}
\usage{
estimateLOQs(
  TE,
  qc = "blanks",
  exclude = "internal.standard",
  sdLOD = 0,
  sdLOQ = 2,
  FUN = NULL,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Estimate LOD/LOQ values from QC samples
}
