% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-testing-functions.R
\name{exampleTracerExperiment}
\alias{exampleTracerExperiment}
\title{Make example TracerExperiment}
\usage{
exampleTracerExperiment(nsamples = 10, nmets = 10, seed = 123, add_qc = FALSE)
}
\arguments{
\item{nmets}{}
}
\description{
Make example TracerExperiment
}
