% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-difftest-functions.R
\name{getContrastSamples}
\alias{getContrastSamples}
\title{Internal function to get the samples/sample names for a given contrast}
\usage{
getContrastSamples(design, contrast, paired = NULL)
}
\arguments{
\item{paired}{}
}
\description{
Internal function to get the samples/sample names for a given contrast
}
