% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-difftest-functions.R
\name{getDataSubset}
\alias{getDataSubset}
\title{Internal function to get the data subset for a given contrast}
\usage{
getDataSubset(data, design, contrast)
}
\arguments{
\item{contrast}{}
}
\description{
Internal function to get the data subset for a given contrast
}
