% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-isocorrection-functions.R
\name{getElementCompositions}
\alias{getElementCompositions}
\title{Retrieve element isotope compositions from NIST}
\usage{
getElementCompositions()
}
\description{
Retrieve element isotope compositions from NIST
}
