% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-difftest-functions.R
\name{getSubsets}
\alias{getSubsets}
\title{Internal function to group contrasts that use the same data subset}
\usage{
getSubsets(contrasts)
}
\arguments{
\item{contrasts}{}
}
\description{
Internal function to group contrasts that use the same data subset
}
