% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-plotting-functions.R
\name{ggcircles}
\alias{ggcircles}
\title{Plot circles with ggplot}
\usage{
ggcircles(
  data,
  mapping = aes(x = x, y = y, fill = value),
  sym = TRUE,
  r = 0.5,
  fontsize = 10,
  title_size = 10,
  labelsize = 10,
  labelface = 1,
  labelcolor = "black",
  col_cex = 1,
  colorscale = c("white", "mediumblue"),
  nacolor = "gray90",
  legend = TRUE,
  linewidth = 0.5,
  plot_margin = 1,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Plot circles with ggplot
}
\examples{
data.frame(x = rep(1:3,3), y = rep(c("A","B","C"), each=3), value = runif(9)*100) |> ggcircles()
}
