% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-imputation-functions.R
\name{imputeTE}
\alias{imputeTE}
\title{Imputation/missing value treatment of isotopologue abundance data}
\usage{
imputeTE(
  TE,
  assay = "lod",
  original = "raw",
  nan = NA,
  na = c("subLODs", "linsubLODs", "LOD50", "missForest"),
  qc_LOD = "lod",
  qc_LOQ = "loq",
  exclude = "internal.standard",
  type = "iso",
  ...
)
}
\arguments{
\item{assay}{cleaned data with missing values}

\item{original}{raw values without missing values}

\item{nan}{replacement for NaN values (default = NA)}

\item{na}{replacement for NA values (number, method or function)}

\item{...}{}
}
\description{
Imputation/missing value treatment of isotopologue abundance data
}
\examples{
exampleTracerExperiment(add_qc = TRUE) |> estimateLOQs() |> preprocessLOQs() |> impute(nan = 0, na = "linsubLODs")
}
