% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-isocorrection-functions.R
\name{isoCorr}
\alias{isoCorr}
\title{Wrapper function for IsoCorrector}
\usage{
isoCorr(
  assaydata,
  molecules,
  Molecule = Molecule,
  MSion = MSion,
  tracer = "C",
  tracer_purity = NA,
  mode = "negative",
  tol_error = 0.05,
  highres = TRUE,
  tmpdir = NULL,
  isocomp = NULL,
  allresults = FALSE
)
}
\arguments{
\item{allresults}{}
}
\description{
Wrapper function for IsoCorrector
}
