% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-data-functions.R
\name{isoEnrichment}
\alias{isoEnrichment}
\title{Fractional enrichment}
\usage{
isoEnrichment(TE, assay = "norm", na.rm.iso = TRUE, na.rm = FALSE, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Fractional enrichment
}
