% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-plotting-functions.R
\name{isoplot}
\alias{isoplot}
\title{Isotopologue plotting}
\usage{
isoplot(
  TE,
  assay = "mid",
  mets = NULL,
  summarise_by = NULL,
  dir = NULL,
  filename = NULL,
  title = NULL,
  title_size = NULL,
  height = 10,
  order_by = NULL,
  dev = "png",
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Isotopologue plotting
}
\examples{
exampleTracerExperiment(nmets = 2) |> MID("raw") |> isoplot(title = "Molecule", colorscale = c("white", "darkblue"), nacolor = "gray50", linewidth = 1, col_cex = 0.5)
}
