% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-data-functions.R
\name{makeTracerExperiment}
\alias{makeTracerExperiment}
\title{Make new TracerExperiment}
\usage{
makeTracerExperiment(
  data,
  metabolite = metabolite,
  label = label,
  QC = NULL,
  metData = NULL,
  colData = NULL,
  meta = NULL,
  assay = "raw"
)
}
\arguments{
\item{assay}{}
}
\value{

}
\description{
Make new TracerExperiment
}
