% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-data-functions.R
\name{MID}
\alias{MID}
\title{Mass isotopomer distribution (MID)}
\usage{
MID(TE, assay = "norm", new_assay = "mid", ...)
}
\arguments{
\item{...}{}
}
\description{
Mass isotopomer distribution (MID)
}
\examples{
exampleTracerExperiment() |> MID(assay = "raw") |> sumMets(assay = "mid", new_assay = NULL)
}
