% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-normalization-functions.R
\name{normalizeIsoAssay}
\alias{normalizeIsoAssay}
\title{Normalization of isotopologue abundance data}
\usage{
normalizeIsoAssay(
  isoAssay,
  method = ~IS + SUM,
  isodata = NULL,
  metAssay = NULL,
  fractions = NULL,
  colData = NULL,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Normalization of isotopologue abundance data
}
