% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-testing-functions.R
\name{rmid}
\alias{rmid}
\title{Generate a random matrix of mass isotopomer distributions}
\usage{
rmid(
  nrow = NULL,
  ncol = 6,
  p = NULL,
  colnames = NULL,
  size = 10^3,
  abundance = NULL
)
}
\arguments{
\item{size}{}
}
\description{
Generate a random matrix of mass isotopomer distributions
}
