% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-data-functions.R
\name{subsetMet}
\alias{subsetMet}
\title{Subset isotobologue data by metabolites}
\usage{
subsetMet(data, met = NULL, isodata = NULL)
}
\arguments{
\item{isodata}{}
}
\value{

}
\description{
Subset isotobologue data by metabolites
}
