% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-data-functions.R
\name{sumMets}
\alias{sumMets}
\title{Summarize isotopologues to metabolites}
\usage{
sumMets(TE, assay = "norm", ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Summarize isotopologues to metabolites
}
