% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{TracerExperiment-class}
\alias{TracerExperiment-class}
\alias{TracerExperiment}
\title{TracerExperiment}
\value{

}
\description{
TracerExperiment
}
\section{Slots}{

\describe{
\item{\code{meta}}{List of metadata (optional)}

\item{\code{colData}}{Data.frame with samples as rownames}

\item{\code{metData}}{Data.frame}

\item{\code{isoData}}{Data.frame}

\item{\code{isoAssays}}{List}

\item{\code{qcAssays}}{List}

\item{\code{metAssays}}{List}

\item{\code{results}}{List}
}}

