% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-plotting-functions.R
\name{variancePartition}
\alias{variancePartition}
\title{variancePartition}
\usage{
variancePartition(data, design, formula = NULL, rm.na = FALSE)
}
\arguments{
\item{rm.na}{}
}
\description{
variancePartition
}
