% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-preprocessing-functions.R
\name{clean}
\alias{clean}
\title{Cleaning of data}
\usage{
clean(
  TE,
  assay = NULL,
  split_by = ~1,
  exclude = "internal.standard",
  maxNAfrac_per_group = 0.2,
  type = "iso",
  ...
)
}
\arguments{
\item{maxNAfrac_per_group}{}
}
\value{

}
\description{
Cleaning of data
}
