% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-difftest-functions.R
\name{diffTest}
\alias{diffTest}
\title{Differential abundance testing of isotope labelling data}
\usage{
diffTest(
  TE,
  contrasts,
  formula = NULL,
  method = "ttest",
  type = "iso",
  assay = "norm",
  logged = FALSE,
  p.adj.method = "holm",
  conf = TRUE,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Differential abundance testing of isotope labelling data
}
