% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-difftest-functions.R
\name{diffTest}
\alias{diffTest}
\title{Differential abundance testing of isotope labelling data}
\usage{
diffTest(
  TE,
  contrasts,
  formula = NULL,
  method = "ttest",
  type = "iso",
  assay = "norm",
  assay_ref = "clean",
  logged = FALSE,
  p.adj.method = "fdr",
  conf = TRUE,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Differential abundance testing of isotope labelling data
}
