% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-data-functions.R
\name{exampleTracerExperiment}
\alias{exampleTracerExperiment}
\title{Make example TracerExperiment}
\usage{
exampleTracerExperiment(nsamples = 10, nmets = 10)
}
\arguments{
\item{nmets}{}
}
\value{

}
\description{
Make example TracerExperiment
}
