% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-data-functions.R
\name{isoEnrichment}
\alias{isoEnrichment}
\title{Fractional enrichment}
\usage{
isoEnrichment(
  TE,
  assay = "mid",
  new_assay = "frac",
  na.rm = TRUE,
  na.rm.iso = TRUE,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Fractional enrichment
}
