% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-data-functions.R
\name{MID}
\alias{MID}
\title{Mass isotopomer distribution (MID)}
\usage{
MID(TE, assay = "norm", max_na = 0, max_imp = 0.25, remove_imp = TRUE, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Mass isotopomer distribution (MID)
}
