% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-preprocessing-functions.R
\name{preprocessLOQs}
\alias{preprocessLOQs}
\title{Preprocessing based on LOD/LOQ estimates}
\usage{
preprocessLOQs(
  TE,
  assay = "raw",
  new_assay = "lod",
  thres_LOQ = 1.5,
  qc_LOD = "lod",
  qc_LOQ = "loq",
  below_LOQ = NA,
  below_LOD = NaN,
  exclude = "internal.standard",
  ...
)
}
\arguments{
\item{TE}{TracerExperiment}

\item{...}{}
}
\description{
Preprocessing based on LOD/LOQ estimates
}
\examples{
exampleTracerExperiment(add_qc = TRUE) |> estimateLOQs() |> preprocessLOQs()
}
