% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-testing-functions.R
\name{rmolecule}
\alias{rmolecule}
\title{Generate random molecule formulas with n C atoms}
\usage{
rmolecule(n = NULL, ...)
}
\arguments{
\item{...}{}
}
\description{
Generate random molecule formulas with n C atoms
}
