% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\name{subset.TracerExperiment}
\alias{subset.TracerExperiment}
\title{Subset TracerExperiment}
\usage{
subset.TracerExperiment(x, factor, type = "col", ...)
}
\arguments{
\item{...}{}
}
\description{
Subset TracerExperiment
}
