% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-data-functions.R
\name{sumMets}
\alias{sumMets}
\title{Summarize isotopologues to metabolites}
\usage{
sumMets(TE, assay = "norm", max_na = 0.1, max_imp = 0.25, ...)
}
\arguments{
\item{...}{}
}
\description{
Summarize isotopologues to metabolites
}
