% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-data-functions.R
\name{sumMets}
\alias{sumMets}
\title{Summarize isotopologues to metabolites}
\usage{
sumMets(
  TE,
  assay = "norm",
  new_assay = "",
  thres_LOQ = 1.5,
  qc_LOQ = "loq",
  max_nafrac_per_met = 0.25,
  max_nafrac_per_group = 0.1,
  min_rep_per_group = 2,
  min_groupfrac_per_iso = 0.8,
  split_by = ~1,
  exclude = "internal.standard",
  sum_qc = NULL,
  na_iso.rm = TRUE,
  na.rm = TRUE,
  ...
)
}
\arguments{
\item{TE}{TracerExperiment}

\item{assay}{norm}

\item{thres_LOQ}{LOQ threshold (values below are set to NA)}

\item{max_nafrac_per_met}{maximum allowed fraction of isotopologue NA values per metabolite}

\item{max_nafrac_per_group}{maximum allowed fraction of NA values per group}

\item{min_rep_per_group}{minimum required number of non-NA values per group}

\item{min_groupfrac_per_iso}{minimum required number of non-NA groups per isotopologue}

\item{split_by}{factor(s) from colData to define groups}

\item{...}{}
}
\description{
Summarize isotopologues to metabolites
}
\examples{
exampleTracerExperiment() |> sumMets(assay = "raw", qc_LOQ = NULL)
exampleTracerExperiment(add_qc = TRUE) |> estimateLOQs() |> preprocessLOQs() |> sumMets(assay = "lod", sum_qc = TRUE)
}
